generalised reference handling in backend

+++ lib/model/Document.php
@@ -35,7 +35,11 @@ class Document extends BaseDocument {
   }
   
   public function delete($con = null) {
+    if(ReferencePeer::hasReference($this)) {
+      throw new Exception("Exception in ".__METHOD__.": tried removing an instance from the database even though it is still referenced.");
+    }
     TagPeer::deleteTagsForObject($this);
+    ReferencePeer::removeReferences($this);
     return parent::delete($con);
   }
